moved TrajectorySegment classes into their own header file
